refactor packages and folders
* ScaleOutSimulator under planner/simulator/scaleout
* ScaleOutSimulation under planner/simulation/scaleout
* factory packages under planner/simulation, planner/simulator
* minkapi typeinfo moved to api/minkapi/typeinfo
