+++ api/minkapi/typeinfo/typeinfo.go
@@ -10,6 +10,7 @@ import (
 	"slices"
 
 	commonconstants "github.com/gardener/scaling-advisor/api/common/constants"
+
 	appsv1 "k8s.io/api/apps/v1"
 	coordinationv1 "k8s.io/api/coordination/v1"
 	corev1 "k8s.io/api/core/v1"
