+++ minkapi/view/base_test.go
@@ -12,9 +12,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/gardener/scaling-advisor/api/minkapi/typeinfo"
-
 	"github.com/gardener/scaling-advisor/api/minkapi"
+	"github.com/gardener/scaling-advisor/api/minkapi/typeinfo"
 	"github.com/gardener/scaling-advisor/common/objutil"
 	"github.com/gardener/scaling-advisor/common/testutil"
 	corev1 "k8s.io/api/core/v1"
