+++ planner/factories.go
@@ -1,8 +1,8 @@
 package planner
 
 import (
-	"github.com/gardener/scaling-advisor/planner/simulation"
-	"github.com/gardener/scaling-advisor/planner/simulator"
+	simulationfactory "github.com/gardener/scaling-advisor/planner/simulation/factory"
+	simulatorfactory "github.com/gardener/scaling-advisor/planner/simulator/factory"
 	"github.com/gardener/scaling-advisor/planner/weigher"
 
 	plannerapi "github.com/gardener/scaling-advisor/api/planner"
@@ -12,12 +12,12 @@ var (
 	_ plannerapi.ScalingPlannerFactory = (*defaultFactory)(nil)
 )
 
-// NewFactories returns a instance of plannerapi.Factories populated with implementation of factory facades.
+// NewFactories returns an instance of plannerapi.Factories populated with implementation of factory facades.
 func NewFactories() plannerapi.Factories {
 	return plannerapi.Factories{
 		Planner:         &defaultFactory{},
-		Simulator:       simulator.NewFactory(),
-		Simulation:      simulation.NewFactory(),
+		Simulator:       simulatorfactory.New(),
+		Simulation:      simulationfactory.New(),
 		ResourceWeigher: weigher.New(),
 	}
 }
