+++ planner/planner.go
@@ -12,7 +12,6 @@ import (
 	"path/filepath"
 
 	"github.com/gardener/scaling-advisor/planner/scorer"
-	"github.com/gardener/scaling-advisor/planner/util"
 
 	commontypes "github.com/gardener/scaling-advisor/api/common/types"
 	plannerapi "github.com/gardener/scaling-advisor/api/planner"
@@ -48,7 +47,7 @@ func (p *defaultPlanner) Plan(ctx context.Context, req plannerapi.Request) <-cha
 		defer close(responseCh)
 		err = p.doPlan(ctx, &req, responseCh)
 		if err != nil {
-			util.SendErrorResponse(responseCh, req.GetRef(), err)
+			SendErrorResponse(responseCh, req.GetRef(), err)
 		}
 	}()
 	return responseCh
@@ -151,4 +150,14 @@ func validateArgs(args *plannerapi.ScalingPlannerArgs) error {
 		return fmt.Errorf("%w: simulationFactory must be set", plannerapi.ErrCreatePlanner)
 	}
 	return nil
+}
+
+// SendErrorResponse wraps the given error with the sentinel error plannerapi.ErrGenScalingPlan, embeds the wrapped error
+// within a plannerapi.Response and sends the response to the given results channel.
+func SendErrorResponse(resultsCh chan<- plannerapi.Response, requestRef plannerapi.RequestRef, err error) {
+	err = plannerapi.AsGenError(requestRef.ID, requestRef.CorrelationID, err)
+	resultsCh <- plannerapi.Response{
+		ID:    objutil.GenerateName("plan-error"),
+		Error: err,
+	}
 }
