+++ planner/simulator/scaleout/singlenode/singlenode.go
@@ -2,16 +2,15 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package multi
+// Package snms provides implementation and helper routines of a ScaleOutSimulator that performs multiple simulations that scale a single
+// node at a time.
+package singlenode
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"sync"
-	"sync/atomic"
 
-	"github.com/gardener/scaling-advisor/planner/util"
+	"github.com/gardener/scaling-advisor/planner/simulator/scaleout"
 
 	"github.com/gardener/scaling-advisor/api/minkapi"
 	plannerapi "github.com/gardener/scaling-advisor/api/planner"
@@ -22,35 +21,23 @@ import (
 )
 
 var (
-	_ plannerapi.ScaleOutSimulator = (*SimulatorSingleNodeMultiSim)(nil)
+	_ plannerapi.ScaleOutSimulator = (*simulatorMultiSim)(nil)
 )
 
-// SimulatorSingleNodeMultiSim is a Simulator that implements ScaleOutSimulator for the SimulatorStrategySingleNodeMultiSim.
-type SimulatorSingleNodeMultiSim struct {
+// simulatorMultiSim is a Simulator that implements ScaleOutSimulator for the SimulatorStrategySingleNodeMultiSim.
+type simulatorMultiSim struct {
 	viewAccess        minkapi.ViewAccess
 	schedulerLauncher plannerapi.SchedulerLauncher
 	storageMetaAccess plannerapi.StorageMetaAccess
 	nodeScorer        plannerapi.NodeScorer
 	traceDir          string
-	state             simulatorState
+	state             scaleout.RequestState
 	simulatorConfig   plannerapi.SimulatorConfig
 }
 
-type simulatorState struct {
-	requestView          minkapi.View
-	simulationFactory    plannerapi.SimulationFactory
-	request              *plannerapi.Request
-	planResultCh         chan plannerapi.ScaleOutPlanResult
-	simulationViews      []minkapi.View
-	simulationGroups     []plannerapi.ScaleOutSimGroup
-	mu                   sync.Mutex
-	simulationRunCounter atomic.Uint32
-}
-
-// New creates a new plannerapi.ScaleOutSimulator that runs multiple simulations concurrently.
-// This is a factory function that supports type plannerapi.ScaleOutSimulatorFactory.
+// New creates a new plannerapi.ScaleOutSimulator that runs simulations for a single scaled node concurrently.
 func New(args plannerapi.SimulatorArgs) (plannerapi.ScaleOutSimulator, error) {
-	return &SimulatorSingleNodeMultiSim{
+	return &simulatorMultiSim{
 		simulatorConfig:   args.Config,
 		viewAccess:        args.ViewAccess,
 		schedulerLauncher: args.SchedulerLauncher,
@@ -64,79 +51,42 @@ func New(args plannerapi.SimulatorArgs) (plannerapi.ScaleOutSimulator, error) {
 // passes where each group is taken in order and simulations run concurrently to get ScaleOutSimResult's. The NodeScorer
 // is used to get ScaleOutSimGroupPassScores; the winner NodeScore determines the simulation View for the next pass,
 // until there are no more passes possible for the ScaleOutSimGroup. Then the ScaleOutSimGroupCycleResult is produced.
-// If the ScalingAdviceGenerationMode is Incremental, a ScaleOutPlanResult is produced from this one cycle result and
+// If the ScalingAdviceGenerationMode is Incremental, a ScaleOutPlanResult is produced from this one-cycle result and
 // sent on the planResultCh, otherwise the cycle result is stored until all cycles are finished. Following which, a
 // cumulative ScaleOutPlanResult is determined from all ScaleOutSimGroupCycleResult's obtained so far and sent on the planResultCh.
-func (m *SimulatorSingleNodeMultiSim) Simulate(ctx context.Context, request *plannerapi.Request, simulationCreator plannerapi.SimulationFactory) <-chan plannerapi.ScaleOutPlanResult {
-	m.state = simulatorState{
-		request:              request,
-		simulationFactory:    simulationCreator,
-		simulationRunCounter: atomic.Uint32{},
-		planResultCh:         make(chan plannerapi.ScaleOutPlanResult),
-	}
+func (s *simulatorMultiSim) Simulate(ctx context.Context, request *plannerapi.Request, simulationFactory plannerapi.SimulationFactory) <-chan plannerapi.ScaleOutPlanResult {
+	s.state = scaleout.RequestStateWith(request, simulationFactory, s.viewAccess)
 	go func() {
-		defer close(m.state.planResultCh)
-		if err := m.doSimulate(ctx); err != nil {
-			util.SendScaleOutPlanError(m.state.planResultCh, request.GetRef(), err)
+		defer close(s.state.ResultCh)
+		if err := s.doSimulate(ctx); err != nil {
+			scaleout.SendPlanError(s.state.ResultCh, request.GetRef(), err)
 		}
 	}()
-	return m.state.planResultCh
+	return s.state.ResultCh
 }
 
-func (m *SimulatorSingleNodeMultiSim) doSimulate(ctx context.Context) (err error) {
-	log := logr.FromContextOrDiscard(ctx)
-	m.state.requestView, err = m.viewAccess.GetSandboxViewOverDelegate(ctx, "request-"+m.state.request.ID, m.viewAccess.GetBaseView())
+func (s *simulatorMultiSim) doSimulate(ctx context.Context) (err error) {
+	if err = s.state.Initialize(ctx); err != nil {
+		return
+	}
+	s.state.SimulationGroups, err = s.createAndGroupSimulation()
 	if err != nil {
 		return
 	}
-
-	if err = util.PopulateView(ctx, m.state.requestView, &m.state.request.Snapshot); err != nil {
-		err = fmt.Errorf("%w: %w", plannerapi.ErrPopulateRequestView, err)
-		return
-	}
-
-	// Run static PVC<->PV Binding for Immediate VolumeBinding mode. Can be done just once for in the requestView for all simulations
-	if _, err = util.BindClaimsAndVolumesForImmediateMode(ctx, m.state.requestView); err != nil {
-		return
-	}
-
-	err = viewutil.LogDumpObjects(ctx, "requestView", m.state.requestView)
-	if err != nil {
-		log.Info("failed to dump view objects", "view", m.state.requestView.GetName(), "error", err)
-	}
-
-	m.state.simulationGroups, err = m.createAndGroupSimulation()
-	if err != nil {
-		return
-	}
-
-	err = m.runStabilizationCyclesForAllGroups(ctx)
+	err = s.runStabilizationCyclesForAllGroups(ctx)
 	return
 }
 
 // Close closes all the resources of this simulator's state: all simulation minkapi views, resets simulation run counters,
 // clears any ScaleOutSimGroup's, clears the planner Request, etc.
-func (m *SimulatorSingleNodeMultiSim) Close() error {
-	m.state.mu.Lock()
-	defer m.state.mu.Unlock()
-	var errs []error
-	for _, v := range m.state.simulationViews {
-		if err := v.Close(); err != nil {
-			errs = append(errs, err)
-		}
-	}
-	clear(m.state.simulationViews)
-	m.state.simulationRunCounter.Store(0)
-	m.state.simulationFactory = nil
-	clear(m.state.simulationGroups)
-	m.state.request = nil
-	return errors.Join(errs...)
+func (s *simulatorMultiSim) Close() error {
+	return s.state.Reset()
 }
 
-func (m *SimulatorSingleNodeMultiSim) createAndGroupSimulation() ([]plannerapi.ScaleOutSimGroup, error) {
+func (s *simulatorMultiSim) createAndGroupSimulation() ([]plannerapi.ScaleOutSimGroup, error) {
 	var allSimulations []plannerapi.ScaleOutSimulation
 	simCount := 0
-	for _, nodePool := range m.state.request.Constraint.Spec.NodePools {
+	for _, nodePool := range s.state.Request.Constraint.Spec.NodePools {
 		for _, nodeTemplate := range nodePool.NodeTemplates {
 			for _, zone := range nodePool.AvailabilityZones {
 				var (
@@ -146,16 +96,16 @@ func (m *SimulatorSingleNodeMultiSim) createAndGroupSimulation() ([]plannerapi.S
 				simCount++
 				simulationName := fmt.Sprintf("sim-%d_%s_%s_%s", simCount, nodePool.Name, nodeTemplate.Name, zone)
 				simArgs := plannerapi.ScaleOutSimArgs{
-					RunCounter:        &m.state.simulationRunCounter,
+					RunCounter:        s.state.SimRunCounter,
 					AvailabilityZone:  zone,
 					NodePool:          &nodePool,
 					NodeTemplateName:  nodeTemplate.Name,
-					SchedulerLauncher: m.schedulerLauncher,
-					StorageMetaAccess: m.storageMetaAccess,
-					Config:            m.simulatorConfig,
-					TraceDir:          m.traceDir,
+					SchedulerLauncher: s.schedulerLauncher,
+					StorageMetaAccess: s.storageMetaAccess,
+					Config:            s.simulatorConfig,
+					TraceDir:          s.traceDir,
 				}
-				sim, err = m.state.simulationFactory.NewScaleOut(simulationName, simArgs)
+				sim, err = s.state.SimulationFactory.NewScaleOut(simulationName, simArgs)
 				if err != nil {
 					return nil, err
 				}
@@ -163,30 +113,30 @@ func (m *SimulatorSingleNodeMultiSim) createAndGroupSimulation() ([]plannerapi.S
 			}
 		}
 	}
-	return groupSimulations(m.state.request.GetRef(), allSimulations)
+	return scaleout.GroupSimulations(s.state.Request.GetRef(), allSimulations)
 }
 
 // runStabilizationCyclesForAllGroups runs all simulation groups until there is no winner or there are no leftover unscheduled
 // pods or the context is done.
-// If the request AdviceGenerationMode is Incremental, after running stabilization cycles for each group it will obtain
+// If the request AdviceGenerationMode is Incremental, after running stabilization cycles for each group, it will obtain
 // the winning node scores and leftover unscheduled pods to construct a ScaleOutPlanResult and send it over the
 // simulator's result channel.
 // If the request AdviceGenerationMode is AllAtOnce, after running all groups it will obtain all winning node scores and
 // leftover unscheduled pods to construct a ScaleOutPlanResult and send it over the simulator's result channel.
-func (m *SimulatorSingleNodeMultiSim) runStabilizationCyclesForAllGroups(ctx context.Context) (err error) {
+func (s *simulatorMultiSim) runStabilizationCyclesForAllGroups(ctx context.Context) (err error) {
 	var (
 		allWinnerNodeScores     []plannerapi.NodeScore
 		simGroupCycleResult     plannerapi.ScaleOutSimGroupCycleResult
 		allSimGroupCycleResults []plannerapi.ScaleOutSimGroupCycleResult
 		log                     = logr.FromContextOrDiscard(ctx)
 	)
-	simGroupCycleResult.NextGroupPassView = m.state.requestView
-	for groupIndex := 0; groupIndex < len(m.state.simulationGroups); {
-		group := m.state.simulationGroups[groupIndex]
+	simGroupCycleResult.NextGroupPassView = s.state.RequestView()
+	for groupIndex := 0; groupIndex < len(s.state.SimulationGroups); {
+		group := s.state.SimulationGroups[groupIndex]
 		log := log.WithValues("groupIndex", groupIndex, "groupName", group.Name())
 		grpCtx := logr.NewContext(ctx, log)
 		log.V(3).Info("Invoking runStabilizationCycleForGroup")
-		simGroupCycleResult, err = m.runStabilizationCycleForGroup(grpCtx, simGroupCycleResult.NextGroupPassView, group)
+		simGroupCycleResult, err = s.runStabilizationCycleForGroup(grpCtx, simGroupCycleResult.NextGroupPassView, group)
 		if err != nil {
 			err = fmt.Errorf("failed to run all passes for group %q: %w", group.Name(), err)
 			return
@@ -197,9 +147,9 @@ func (m *SimulatorSingleNodeMultiSim) runStabilizationCyclesForAllGroups(ctx con
 			continue
 		}
 		allWinnerNodeScores = append(allWinnerNodeScores, simGroupCycleResult.WinnerNodeScores...)
-		if m.state.request.AdviceGenerationMode.IsIncremental() {
-			log.V(4).Info("Sending ScalingPlanResult", "adviceGenerationMode", m.state.request.AdviceGenerationMode)
-			if err = util.SendScaleOutPlanResult(ctx, m.state.planResultCh, m.state.request, m.state.simulationRunCounter.Load(),
+		if s.state.Request.AdviceGenerationMode.IsIncremental() {
+			log.V(4).Info("Sending ScalingPlanResult", "adviceGenerationMode", s.state.Request.AdviceGenerationMode)
+			if err = scaleout.SendPlanResult(ctx, s.state.ResultCh, s.state.Request, s.state.SimRunCounter.Load(),
 				[]plannerapi.ScaleOutSimGroupCycleResult{simGroupCycleResult}); err != nil {
 				return
 			}
@@ -215,9 +165,9 @@ func (m *SimulatorSingleNodeMultiSim) runStabilizationCyclesForAllGroups(ctx con
 		err = plannerapi.ErrNoScaleOutPlan
 		return
 	}
-	if m.state.request.AdviceGenerationMode.IsAllAtOnce() {
-		log.V(4).Info("Sending ScalingPlanResult", "adviceGenerationMode", m.state.request.AdviceGenerationMode)
-		err = util.SendScaleOutPlanResult(ctx, m.state.planResultCh, m.state.request, m.state.simulationRunCounter.Load(), allSimGroupCycleResults)
+	if s.state.Request.AdviceGenerationMode.IsAllAtOnce() {
+		log.V(4).Info("Sending ScalingPlanResult", "adviceGenerationMode", s.state.Request.AdviceGenerationMode)
+		err = scaleout.SendPlanResult(ctx, s.state.ResultCh, s.state.Request, s.state.SimRunCounter.Load(), allSimGroupCycleResults)
 	}
 	return
 }
@@ -227,7 +177,7 @@ func (m *SimulatorSingleNodeMultiSim) runStabilizationCyclesForAllGroups(ctx con
 //   - the simulation group has stabilized with no scheduled pods for all its child simulations.
 //   - there is no winner node score after running a pass for the group
 //   - the context is done.
-func (m *SimulatorSingleNodeMultiSim) runStabilizationCycleForGroup(ctx context.Context, groupPassView minkapi.View, group plannerapi.ScaleOutSimGroup) (cycleResult plannerapi.ScaleOutSimGroupCycleResult, err error) {
+func (s *simulatorMultiSim) runStabilizationCycleForGroup(ctx context.Context, groupPassView minkapi.View, group plannerapi.ScaleOutSimGroup) (cycleResult plannerapi.ScaleOutSimGroupCycleResult, err error) {
 	var (
 		winningNodeScore *plannerapi.NodeScore
 	)
@@ -242,7 +192,7 @@ func (m *SimulatorSingleNodeMultiSim) runStabilizationCycleForGroup(ctx context.
 			cycleResult.PassNum++
 			log := logr.FromContextOrDiscard(ctx).WithValues("groupRunPassNum", cycleResult.PassNum)
 			passCtx := logr.NewContext(ctx, log)
-			cycleResult.NextGroupPassView, winningNodeScore, err = m.runSinglePassForGroup(passCtx, cycleResult.NextGroupPassView, group)
+			cycleResult.NextGroupPassView, winningNodeScore, err = s.runSinglePassForGroup(passCtx, cycleResult.NextGroupPassView, group)
 			if err != nil {
 				return
 			}
@@ -274,19 +224,19 @@ func (m *SimulatorSingleNodeMultiSim) runStabilizationCycleForGroup(ctx context.
 // invokes the NodeScorer for each valid ScaleOutSimResult to compute the NodeScore and aggregates scores into the ScaleOutSimGroupPassScores - which includes the WinnerScore if any.
 // If there is a WinnerScore among the SimulationRunResults, within the SimulationGroupRunResult, it is returned along with the nextGroupView.
 // If there is no WinnerScore then return nil for both winnerNodeScore and the nextPassView.
-func (m *SimulatorSingleNodeMultiSim) runSinglePassForGroup(ctx context.Context, groupPassView minkapi.View, group plannerapi.ScaleOutSimGroup) (nextGroupPassView minkapi.View, winnerNodeScore *plannerapi.NodeScore, err error) {
+func (s *simulatorMultiSim) runSinglePassForGroup(ctx context.Context, groupPassView minkapi.View, group plannerapi.ScaleOutSimGroup) (nextGroupPassView minkapi.View, winnerNodeScore *plannerapi.NodeScore, err error) {
 	log := logr.FromContextOrDiscard(ctx)
 	var (
 		groupScores plannerapi.ScaleOutSimGroupPassScores
 		winnerView  minkapi.View
 	)
 	simulationRunResults, err := group.Run(ctx, func(ctx context.Context, name string) (minkapi.View, error) {
-		return m.createSandboxView(ctx, name, groupPassView)
+		return s.state.CreateSandboxView(ctx, name, groupPassView)
 	})
 	if err != nil {
 		return
 	}
-	groupScores, winnerView, err = m.processSimulationGroupRunResults(log, group.Name(), simulationRunResults)
+	groupScores, winnerView, err = s.processSimulationGroupRunResults(log, group.Name(), simulationRunResults)
 	if err != nil {
 		return
 	}
@@ -304,18 +254,7 @@ func (m *SimulatorSingleNodeMultiSim) runSinglePassForGroup(ctx context.Context,
 	return
 }
 
-func (m *SimulatorSingleNodeMultiSim) createSandboxView(ctx context.Context, name string, groupPassView minkapi.View) (minkapi.View, error) {
-	m.state.mu.Lock()
-	defer m.state.mu.Unlock()
-	sandboxView, err := m.viewAccess.GetSandboxViewOverDelegate(ctx, name, groupPassView)
-	if err != nil {
-		return nil, err
-	}
-	m.state.simulationViews = append(m.state.simulationViews, sandboxView)
-	return sandboxView, nil
-}
-
-func (m *SimulatorSingleNodeMultiSim) processSimulationGroupRunResults(log logr.Logger, simulationGroupName string, simulationRunResults []plannerapi.ScaleOutSimResult) (simGroupRunScores plannerapi.ScaleOutSimGroupPassScores, winningView minkapi.View, err error) {
+func (s *simulatorMultiSim) processSimulationGroupRunResults(log logr.Logger, simulationGroupName string, simulationRunResults []plannerapi.ScaleOutSimResult) (simGroupRunScores plannerapi.ScaleOutSimGroupPassScores, winningView minkapi.View, err error) {
 	var nodeScore plannerapi.NodeScore
 
 	for _, sr := range simulationRunResults {
@@ -323,7 +262,7 @@ func (m *SimulatorSingleNodeMultiSim) processSimulationGroupRunResults(log logr.
 			log.V(2).Info("No ScaledNodePodAssignments for simulation, skipping NodeScoring", "simulationName", sr.Name, "simulatedNodePlacement", sr.ScaledNodePlacements[0])
 			continue
 		}
-		nodeScore, err = m.nodeScorer.Compute(mapSimulationResultToNodeScoreArgs(sr))
+		nodeScore, err = s.nodeScorer.Compute(mapSimulationResultToNodeScoreArgs(sr))
 		if err != nil {
 			err = fmt.Errorf("%w: node scoring failed for simulation %q of group %q: %w", plannerapi.ErrComputeNodeScore, sr.Name, simulationGroupName, err)
 			return
@@ -331,7 +270,7 @@ func (m *SimulatorSingleNodeMultiSim) processSimulationGroupRunResults(log logr.
 		simGroupRunScores.AllScores = append(simGroupRunScores.AllScores, nodeScore)
 	}
 	if len(simGroupRunScores.AllScores) > 0 {
-		simGroupRunScores.WinnerScore, err = m.nodeScorer.Select(simGroupRunScores.AllScores)
+		simGroupRunScores.WinnerScore, err = s.nodeScorer.Select(simGroupRunScores.AllScores)
 		if err != nil {
 			err = fmt.Errorf("%w: node score selection failed for group %q: %w", plannerapi.ErrSelectNodeScore, simulationGroupName, err)
 			return
