+++ planner/simulator/simulator_test.go
@@ -1,4 +1,4 @@
-package util
+package simulator
 
 import (
 	"context"
